backend/pyproject.toml
[project]
name = "ai-chat-backend"
version = "0.1.0"
description = "AI Chat Web App - Django Backend with LangGraph"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.4",
    "gunicorn>=22.0",
    "uvicorn[standard]>=0.30",
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "psycopg[binary]>=3.2",
    "pgvector>=0.3",
    "redis>=5.0",
    "channels>=4.0",
    "channels-redis>=4.2",
    "langgraph-checkpoint-postgres>=2.0",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "tavily-python>=0.5",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests (requires DB + real APIs)",
    "eval: Agent evaluations (requires OpenAI + Tavily keys)",
    "slow: Slow tests (deep research, multi-turn)",
]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["config", "chat", "documents"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
